Split ExpressionBuilder* and Context into their own files
