+++ Cql/Cql.Compiler/ExpressionBuilderContext.Scopes.cs
